Emit "blu" for i_bltu instead of "bltu".
I'm unable to find documentation of bltu. GAS does not accept bltu; Sun's as(1) in Solaris 10 does, and appears to output the same byte sequence as it does for blu. Since it appears that bltu is a mnemonic for blu, I'm just modifying this to also output blu, which should suit both assemblers.

svn path=/trunk/; revision=2826