Moved `__WRONG_*` markings out of the startup headers.
I'm putting these under `api/` instead, because this is the place where (in the Makefile) we express which APIs are present. So it also seems a natural place to express (via the __WRONG macros) which subsets of APIs are not present.

For sensible OSes, this should leave startup/ portable. All that should remain there are whatever OS-specific flags bring in conformance for each API. For example an OS might respond to `_ANSI_SOURCE` being defines and avoid presenting any non-ANSI interfaces to programs which include its headers.

svn path=/trunk/; revision=2827