Seperated binasm output from symbolic assembly text, inverting the -S option. This makes -S output binasm, and to do so instead of symbolic assembly text rather than in addition.
This changes the parameteres required to be passed to mips and alpha trans; these are now just a .t and .s file, or (with -S present) .t and .T/.G. tcc(1) will need updating accordingly, which I shall do in a seperate commit.

The error-checking (and testing) here is minimal; this is just a moving-stuff-around sort of sketch to get things roughly in the right areas.

svn path=/trunk/; revision=2830