Consolidated a few hacked includes for glibc and eglibc. Despite being in `bits/` these headers have *hacks* which appear to be portable between various architectures.
Since the hacks are not architecture dependent, they can live under `common/`, which is desirable to reduce repetition and to ease adding new architectures where support for the libc already exists due to an existing architecture.

Note that it is not relevant if the libc headers' contents are or are not architecture-dependent.

svn path=/trunk/; revision=2842