Reworked startup file creation.
Previously the startup (and endup and tokdef) files were created outside of tempdir. This was because the location of the directory isn't neccessarily known until the environments are parsed, and the startup files may need to be created earlier due to various command line options such as `-D`.

Here I've removed the `temporary_dir` variable, so that the directory location is given directly by the `PREFIX_TMP` environment variable. The `-temp` option (should) simply override this. So now the directory is created in one of two places: either as required on-the-fly when writing to a startup file, or after options are parsed.

I don't really like the on-the-fly creation, but it's the best approach I can think of for now.

svn path=/trunk/; revision=2847