Moved up `isxyz` macros to `PART_CFLAGS` for each trans's Makefile.
These are used to identify which trans is being built, which is only relevant for building PARTS (i.e. shared code under common/ which has trans-specific bits and pieces). This isn't the same as identifying which CPU is being targetted, except insofar as each trans targets a particular family. I'll rename these shortly to make that clearer.

svn path=/trunk/; revision=2865