Workaround for trailing whitespace for make(1) on OpenBSD 5.3.
The line here:
{{{
PREFIX_LIBEXEC?= ${PREFIX}/libexec # should not be in $PATH
}}}

assigns the value "xyz/libexec" for make(1) on OpenBSD 5.2, but "xyz/libexec " (note the trailing space) on OpenBSD 5.3. This appears to be true for trailing whitespace in general; the comment has no effect.

Here I've moved the comment elsewhere so that the whitespace can also go.

svn path=/trunk/; revision=2904