Replace tests against `... != ""` with `!empty(...)` instead.
I was sure `empty()` failed in some context for a certian BSD make dialect, which is why I switched to `... != ""` here in the first place. However it it seems apparently not. Perhaps I was confusing that with the situation where a variable is not defined. That's tested for seperately here, by the surrounding `defined(...)` where neccessary.

Tested this on pmake, bmake, and NetBSD, OpenBSD and FreeBSD's native makes.

svn path=/trunk/; revision=2907