Added a pragma to permitinterchanging function and object pointers for POSIX.
This is a requirement POSIX places, whereby the representations for both are considered equal, and conversions between each are permitted. The effect here is that under `-Yposix` tcc silently permits this behaviour, and otherwise is able to complain about it.

Note currently such diagnositics are not generated under `-Xp`, which is the default.

svn path=/trunk/; revision=2913