Workaround for invalid assembly generated for a negative `min`.
This would output:
{{{
jmp *.L55--40(,%edx,4)
}}}

where `min` is negative. I believe that should be `.L55 + 40` instead, and so this commit adjusts to cater for that.

I'm unsure that a negative `min` ever makes sense here in the first place. For the moment though, this is just a quick fix.

Thanks for David Holland for thoughts on this.

svn path=/trunk/; revision=2917