Moved out API building.
This is a start on restructuring the OS-dependent things. My goal here is to make that a bit smaller, and generally easier to approach. To start with, I'm splitting off a few unrelated parts so that I may then reduce down each item to something more manageable.

In the case of APIs here, I'm taking the premise that if an OS professes to implement a certian API, it will do so for all architectures it supports. If it doesn't, that's actually something we can deal with anyway - but I don't think it warrants a per-architecture level of subdirectory. That's not done in this commit; it's just an example of how the layout here might be simplified, now that `/api` is an independent part.

svn path=/trunk/; revision=2924