Restructured libc directory layout.
This is a start on tidying things up a bit for `/api`, now that it is separated from `/osdep`. In this commit I have flattened down what was the `machines/` heirachy to lose its per-architecture layer, as this is unneccessary for the APIs presented as interfaces by each libc implementation.

There are architecture-specific left-overs, which for the moment live under `arch/`. These can be (mostly) flattened out to `../include` in a later commit. Things like `stdarg.h` may be better to be kept separate, though, depending how it turns out.

svn path=/trunk/; revision=2926