Workaround for include paths on Debian-like systems.
I am not a fan of this being here, but other solutions are not immediately apparent. This also does not cater for paths for the CRT files, which must be set in tcc environments. Ideally I would like both these things (somehow) set in the same place, rather than spreading OS-specific information across the source tree. Suggestions welcome.

svn path=/trunk/; revision=2933