Converted `only_lengthen_ops` to a runtime flag, `OPTIM_SHORTEN`.
This is the inverse of the origional meaning, since the effect of setting an OPTIM_ flag is to cause an optimisation to occur - in this case to shorten integer sizes where possible.

Architecture-specific environments have been added where neccessary to maintain the origional `config.h` behaviour per platform, namely to disable this for 80x86, alpha and mips.

svn path=/trunk/; revision=2941