Converted `promote_pars` to a runtime flag.
This is not exposed as a CLI option, as I don't think it makes sense for it to be configurable. The reason for this is that whether promotion is desirable or not depends largely on the word size for a given particular architecture, at least as far as I understand. The installer itself is the only entity with that knowledge.

svn path=/trunk/; revision=2954