Removed `DO_SCHEDULE` and `do_scheduling`. This drops the instruction scheduling ability for the Alpha installer.
Technically this ability wasn't ever present in this code; all this variable did was to have instructions accumulate into a string, and then whenever a certian number have been stored, to pass them to the instruction scheduler for output.

However, the instruction scheduler does not actually exist in TenDRA's code. This was provided by:
{{{
#if DO_SCHEDULE
#include "scheduler.h"
#endif
}}}

which I presume is either an external product, or is provided by OSF1. I have been unable to find this header - and in order to change the `DO_SCHEDULE` macro to a runtime condition it would need to be present, as there are several types and symbols which the code depends on.

So, since all this code really does is produce strings, I am removing this.

If we do ever want to replace this functionality, I think this would be better done inside commonly-shared opcode output functions themselves, rather than spreading conditions throughout the code at every point those functions are called.


svn path=/trunk/; revision=2960