ABI is now a command line option, `-A`.
Most of the ABI names here are a bit silly. They need looking through and naming properly.

There are lots of variables scattered around which set various installer-specific options; many of these belong under ABI selection. This is to be done in a separate pass. My focus with this commit is just to get the `-A` flag in place, and to remove any remaining OS-specific macros.

svn path=/trunk/; revision=2994