Renamed `/producers-dra` to `/tdfc2`, for much the same reasons as [2999].
This provides both the C and C++ producers, which (currently) are two seperate executables, named `tdfc2` and `tcpplus` respectively. I am undecided, but it is possible that I may eventually merge these, with a runtime option to select the language. Probably not, though.

I'm naming this top-level "tdfc2" rather than "tcpplus" because that helps distinguish it from the origional (C-only) tdfc, which we don't have.

svn path=/trunk/; revision=3000