Removed the `COPYRIGHT` feature.
This feature was present in a milder form in the 4.1.2 release of Lexi, by way of copying through the first comment of the `.lxi` file to the generated code. Kevin and I converted that to take a filename instead, because now we have multiple copyrights and the magic comment behaviour (although lovely in its simplicity) is a bit obtuse.

Now I'm removing it entirely and simply generating this, instead:

{{{
/*
 * Automatically generated by lexi version 2.0
 * Copyright terms for the input source also apply to this generated code.
 */
}}}

My rationalle is simply that I don't think this is worth spending this many lines of code for.

svn path=/trunk/; revision=3018