Refactoring; rewrite of the local names as a simple linked list rather than a trie.
I have no idea why this was so elaborate in the first place; especially the code generation iterated from each leaf to the root in order to reconstruct names for node paths; why not just store the names in the first place?

Nodes are prepended to this list, rather than appended. That means the arguments are traversed and output in reverse order. I think this is okay, as order does not matter here, as long as it is consistent.

svn path=/trunk/; revision=3082