Split XPG4 into subsections, reflecting how they are grouped in the document.
This was done by retrospectively grouping based on the XPG4v2 document (which is all I have available), based on the differences documented from XPG4. XPG4v2 also adds a "ux" section, which is ignored here.

All the interfaces present here are from the XSH section of XPG4, so these are named `xsh4-*` rather than `xpg4`.

svn path=/trunk/; revision=3182