First steps at defining a language for expressing execution character sets. For the moment I'm writing this as a seperate tool until it's ready for use. I specifically don't want this to be complex enough to warrant a proper lexer and parser, so this is largely an experiment to see if that's feasible.
I'm using a language here so that arbitary character sets may be defined at runtime, rather than building them into `tdfc2`. An alternative approach would be to have an external tool generate a simple format file (simple to the extreme of just listing every character's value), and have `tdfc2` read that format only. However I think the reading of that format is around as complex as generating it in the first place and I'd rather have one format than two.

svn path=/branches/tdfc2-charset/; revision=3225