Added calls to load and perform character set mapping.
Finding the right spot to do this was tricky. This is implemented in the handling for string and character literals.

`new_string_lit()` despite its name deals with both string and character literals (and is commented as such). Prior to that, there's a mapping from the source character set to ASCII, and `new_string_lit` is given those ASCII strings. So there I can map from source to ASCII, and then later on use the existing mapping from ASCII to TDF bitstreams (in `output/`), and then from TDF to the host encoding (in the installers).

`new_string_lit` is only passed characters 0..255 inclusive though (i.e. already mapped from unicode) but before escape sequences such as `\t` are translated. This mapping occurs after those escape sequences are translated, and then the result of that translation is mapped.

Note that this cannot be done in the caller of `new_string_lit`, because `new_string_lit` can return a "shared" string literal (i.e. the tail of some other string literal).

svn path=/branches/tdfc2-charset/; revision=3234