A start on a common `main()` to be shared by all architectures.
The general idea is to provide one point for drivers to hook on to various things. It'll take a while to move everything under here, which I'll do incrementally.

One thing that doesn't fit under this model, are the `<local/*.h>` headers for macros and types included from the `common/` code.

svn path=/trunk/; revision=3266