Off by one.
I'm not using `memmove()` here because for the first instance `res->exp` (and therefore `i`) may be 0. For the second instance that can never occur (since the exponent is always >= 0 at that point), but I think the code should match for both cases.

Initially spotted by Tobias, and again by Vigud.

svn path=/trunk/; revision=3298