This commit brings across headers from `diag/`, with their structure and layout left as-is.
Some background:

The general goal here is to have all diagnostic output formats linked simultaneously, so that they can be selected at runtime. The trouble is, some use the `diag/` interface and some use the `newdiag/` interface. There are several possible ways to deal with this, and none are as attractive as doing the "right thing" in updating all the code to use the same interface, and then deleting the unused one.

The task of how to go about updating code to use the `newdiag/` interface is not straightforward. There is quite a lot of code (even within one architecture) which uses the older `diag/` interface, and the trouble is, when updating those from `diag/` to `newdiag/`, I can't just test that one thing at a time because switching datastructures is an all-or-nothing approach.

So, this commit starts an attempt to address that problem.

The plan for this branch is as follows:

 1. Add these `diag/` fields to their corresponding `newdiags/` structs;
 2. Update code to use these "stop-gap" fields within the `newdiag/` structs and switch everything to build using `newdiag/`;
 3. Individually replace each instance of use of these imported `diag/` fields to use the `newdiag/` layout instead;
 4. Remove the `diag/` "stop-gap" fields and `diag/` entirely.

Each item here is prefixed `reform_` for ease of identification at the points where it is to be used.

Other than field names changing, there should be very little (i.e. almost no) change to code when the `diag/` fields from these headers are used in-situ inside the `newdiag/` structs.

svn path=/branches/trans-diagreform/; revision=3365