Enable handling for `diagnose_tag` under `-DDIAGREFORM`.
This is present for `diag/`. My understanding is that it is of historic significance only; I can find no mention of `~diagnose` in the TDF Diagnostic spec., nor does it seem to appear in the tdfc2 producer. My best guess is that this tag used to exist, but is no longer present.

I'm enabling it just because `diag/` does use it, and my aim here is to keep compatibility with the behaviour there.

svn path=/branches/trans-diagreform/; revision=3369