Added call out to `INSPECT_FILENAME()`.
This is part of the `diag/`-compatible interface for `newdiag/`. It seems installers use this to build lists of files, which I believe to be unneccessary for the `newdiag/` interface. So this is added here as part of the stop-gap scheme for merging the two interfaces.

It will probably eventually be removed again when those lists of files are also removed, and that code is updated to use the `newdiag/` scheme for handling files.

svn path=/branches/trans-diagreform/; revision=3371