Rework binary encoding for `SORT_tdfstring`. This was was believed to be incorrect, and segfaulted on the `multi_string` in the following test case:
{{{
( make_tokdec ~signed_int
  -
  variety )

( make_id_tagdec fish
  -
  -
  proc )

( make_id_tagdef fish
  -
  ( make_nof_int
    ~signed_int
    ( multi_string
      123
      456
      789 | ) ) )
}}}

I'm not sure how the origional code came to be, or if perhaps it catered to a situation we don't understand. Per the TDF spec, a TDFSTRING ought to consist of the number of bits per character, followed by a repetition of the characters. This seems to be almost what the code does, except the tree traversal is slightly off in depth of nodes.

As I understand, the `if (q->encoding == -1)` case caters to a TDFSTRING consisting entirely of `SORT_small_tdfint`, and the opposing case is for a TDFSTRING consisting entirely of `SORT_tdfint`. I do not believe a mix can ever occur, and so recurring to deal with the node type here is unneccessary.

Likewise there can never be anything other than a `SORT_repeat` for the body of the string, and so the tree traversal for that is hardcoded, too.

I am still unclear as to whether octal digits are possible here (as the encoding has origionally), and if so, how to produce them.

Debugged by myself and edk.

svn path=/trunk/; revision=3384