A start on attempting to simplify the paths for tspec ...by first making them more complicated.
My goal here is to keep the generated headers for subsets hidden from user programs; this commit moves those to their own directory, so that `#include` does not see them. However the API interface headers (such as `<stdio.h>`) do need to include them, and they do so by relative paths. This is not done very well, but it'll do for now.

svn path=/trunk/; revision=3451