Moved out `.~bitfield_is_signed`, with an associated ABI token, `.~abi_bitfield_is_signed`.
At a cursory glance, this does not seem to be used by either the producers or installers. Perhaps it was used historically.

I might remove it - it's part of the token register for the C LPI tokens, which are defined ad-hoc rather than being formally part of the TDF spec. So I'd be okay with removing it, since the C LPI is specific to this producer anyway.

svn path=/trunk/; revision=3483