Move `K-ppc` under `arch/power`; what remains is identical to `power/`, and so that is removed.
How to automate selection of `K-ppc` vs. `K-power` remains to be determined. My instinct is for this to be similar to selecting between any family of CPU. For example, I'd like tcc to default to calling `trans.x86 -K5` for 586 where available, rather than the lowest common denominator. We don't currently have a mechanism to express that; possibly it might be another `tendra.makedefs.mk` variable.

svn path=/trunk/; revision=3629