Diagnostics are stored per `dec`; the single field there is the entry point into the diagnostics datastructures. This commit splits that into two fields, `.dg_name` for with `NEWDIAGS` and `.diag_info` for without. The rest of this commit is just to deal with the additional name.
This is a step towards being able to compile with both present simultaneously. For the moment these are mutually exclusive, but my intention is that these `#ifdef` conditionals can be removed, and both fields can be present together. Then (eventually) either or both can be populated, depending on what tokens are present in the TDF capsule.

svn path=/trunk/; revision=3727