`trans.x86` has a special case for optimising the following, such that the jump is removed: {{{ jmp .L66 .L66: }}}
The way that redundant jump is optimised out is implemented really horribly; the output routines optimise away the redundant jump by seeking back to the file offset just before the jump was output. I either want to remove that code, or achieve the same effect in a nicer way.

I think the proper thing to do here is to remove this optimisation entirely, and then implement it in a more sensible place instead (i.e. operating on the TDF, before output). But I don't want to go down that particular rabbit hole at the moment.

For now I'd rather have less optimial assembly generated with more maintainable source. So I'm removing this optimisation (in part because `trans.x86` is the only architecture which needs to keep track of the output position in that way, which it does on every newline, but also because it's plain horrible), and if it returns, it can hopefully live in a more appropriate place, operating on TDF before output.

svn path=/trunk/; revision=3732