The various architectures print assembly by different mechanisms; `fputs` and friends, `fprintf`, "out" functions, and a few other bits and pieces. And of course some of code mixes various ways. The "out" functions in particular cause trouble, because they're defined slightly differently for both `sparc` and `x86`, and this gives special cases for the `common/` code which calls them.
My aim here is to bring some consistency to the way in which assembly is formatted out (i.e. to pretty it up), and a first step to that is to have everything use the same mechanism for printing text. This commit is an attempt to convert all output to use the printf-style `asm_` functions. My intention is that these functions can be fitted out with more elaborate handling for formatting and such, now that all the code goes via that interface.

There are many cases where slightly the wrong format specifier is used for the type of its argument. Some of these may be a result of my conversion of e.g. the "out" functions en-masse, but most of these seem to be due to incorrect types for the origional printf format strings, where present. I have preserved these errors, as I am not attempting to fix those in this commit; this commit is just the normalisation for all architectures to use the same interface.

svn path=/trunk/; revision=3804