Remove the `TDF_OUTPUT` macro.
This allowed tdfc2 to be built "without" support for TDF output. What actually happened was that a valid capsule is generated, but contains no information. This is just the same as the runtime option to elide TDF otuput. I don't see any reason to have this as a compile-time option.

svn path=/trunk/; revision=3895