Type consistency for `stab_file()`.
I see no reason why an index into a list of files need be `long`. At least some architectures use this as an fd number; however I am not sure if all do, or if that relates to some other field. So for now I'm making them all `long`, just in case. This is probably overkill.

svn path=/trunk/; revision=4013