Switch from `<literallayout>` to providing newlines by a processing instruction, `<?br?>` instead.
This is an extension to Docbook, albeit without introducing a new entity. The only change to the DTD here is to remove `<literallayout>` added in [4040].

I'm providing a `&br;` entity for this PI, so that it's a little saner to type. The intended use cases (also updated in this commit) are for newlines inside `<td>` elements, and newlines inside `<title>` elements. The latter is rendered out conditionally as a containing block, such that the content is aligned agaist section numbers.

svn path=/trunk/; revision=4043