Only set `FS_POSIX` under `__TenDRA__`.
I managed to remove a few POSIX dependencies with no change of functionality. but a couple remain where the features are useful - like having the proper paths for files in diagnostics. I'd like to keep those, but they're not important for bootstrap. So, I'm providing those "for real", but not during bootstrap.

svn path=/trunk/; revision=4136