Some consistency for the `make_code()` entry points under `TDF_DIAG4`.
The diag4 diagnostic tokens are handled by an "extra" function (presumably added poc-hoc) which deals with those tokens before calling the bulk of the handling. This was elided under the absence of `TDF_DIAG4`.

The change here is to always have that wrapper present, and to select on diags internally. This keeps the same interface for both cases, such that they can be centralised in the future.

svn path=/trunk/; revision=4163