Remove `FBASE == 10`.
As far as I understand, this was for floats represented as sequenced of base 10 digits encoded as ASCII characters. This is not the same as sequences of base n digits encoded as numerical values, which is what we have for `FBASE == 65536`. So I think the remaining code still does support arbitary base encodings; what's removed here is just the special case handling for ASCII characters.

svn path=/trunk/; revision=4295