Populate `restrict` placeholders for forwards compatibility.
This is an attempt to keep clutter down, after trying the more orthodox approach of moving things into subsets, so they may be overridden.

C99 onwards qualify several things with `restrict`, but are otherwise identical to their specifications for earlier standards. Rather than fragmenting these into subsets and overriding those to introduce `restrict`, a definition is made conditionally in `master.ts`. This helps keep related functions together, especially for headers which are already complex.

The downside is of course that it seems inappropriate to have `~restrict` appear in C89 headers, even if it expands out to nothing. But the benefit is that it makes both C89 and C99 headers an awful lot simpler.

An alternate solution to the same problem would be to implement C99's headers as standalone things, without inheriting from earlier standards. I am not willing to do this; part of tspec's purpose is to convey inheritence from one standard to the next.

svn path=/trunk/; revision=4321