The hosted and freestanding APIs for C99.
Stefan undertook an implementation of most of C99's API for ten15.org's repository. My approach differs somewhat; his was defined standalone rather than from previous standards by `+IMPLEMENT` or `+USE` to derive from them.

Earlier I'd added C95, and so I also want to derive from that. So I have not imported Stefan's work; instead I've implemented this from scratch so that I can code it in terms of inheritence from C89 and C95 where appropriate.

This API is specified per ISO/IEC 9899:1999+TC1+TC2+TC3, but arguably the Technical Corrigenda ought to appear as seperate specifications instead. I am unsure as to whether we ought to expose the differences in tspec. For the moment, I'm taking the stance that the TC are intended to correct mistakes and omissions, and therefore supersede previous revisions. In other words, that they're part of the same standard, rather than being something worth encoding in and of themselves.

svn path=/trunk/; revision=4323