Remove `EXTERN_DIAG`.
This was defined as `extern` for files which use symbols, and defined to expand out to nothing for files which provide those symbols.

I've seen this technique used occasionally, and I think it's a cute idea with the intention of avoiding repetition. However it's rarely done, and nothing else uses it. So I'm removing it here because it's confusing.

svn path=/trunk/; revision=4391