Move out `.frame`, independently from the `al_val` pseudo-union.
The `al>val` struct seems to be intended as a union. But the `.frame` field seems to be used simultaneously to the other fields, and I'm guessing it was added after this struct was made. So I'm moving it out, such that the remaining struct can become a union instead.

svn path=/trunk/; revision=4398