This is a fix for what seems to be an out of bounds off-by-one array access.
The `BITSTREAM` linked list (which is used for encoding TDF) operates as a queue of buckets, with upto `CHUNK_SIZE` bits per node. The idea is to output as much as possible upto an even byte's worth of bits, leaving the remaining bits for the next pass.

So this means the "left over" bits will always be less than a byte, and (since `CHUNK_SIZE` divides exactly into a byte) there will be no bits remaining when an entire chunk is written out.

{{{
bs->text[0] = bs->text[bs->bytes];
}}}

The error is that this is called when `bs->bytes == CHUNK_SIZE`. However there would be no remaining bits for that case anyway, and so my fix is to simply elide the assignment in that case.

svn path=/trunk/; revision=4449