THIS IS CURRENTLY WIP.  I'm asking for comment on the makefile and env changes.
One thing I see that should be improved in tendra's makefiles
and env files is to separate the traditional prefix part of the
library and include file paths "/usr", "/develop" etc from the
downstream parts "./lib", "./include" which are considered
relocatable and underneath by (for example) configure.  For now
I've followed tendra's scheme in the builtin musl, which adds
a spurious "/usr" under obj.*/tcc/sys/libc.  It makes sense to
break up tendra's idea of prefix into two parts sometime in the
future.

First shot at tendra side changes to allow tendra to ship with
(or with the ability to generate) a 32-bit musl libc 1.1.5 with
prefix osdep/libc/musl-1.1.5/usr.

I've added (for now) a script to rebuild the musl libc in
shared/bin/build-musl-libc.sh.  It's long but should be close
to as solid as it's possible to get.  I'm sure this will move
in the final code.

This invocation configures things so:

bmake \
      TARGETARCH=x32_64 \
      LIBCPREFIX=/home/arty/dev/tendra/tendra/osdep/libc/musl-1.1.5 \
      LIBC_VER=MUSL_1_1_5 USE_LIBC=builtin

I've made an effort not to disturb the normal operation of the
makefiles here so I've introduced a new variable, USE_LIBC into
the makefiles that carries either 'builtin' or 'system' (default
system).

The makefiles in osdep/machines/musl/x32_64/env and x86_32 now
copy the env subdir from ../common/${USE_LIBC}-template on the
.BEGIN rule.  It's gross but I admittedly didn't know what rules
would be invoked in there.

I haven't included the built musl libc in this commit so as not
to pollute the tree at this point.  A buildable container of
some kind is forthcoming that duplicates that part of the build
fully.