Add chapter files as dependencies for TDF documentation makefiles
Modifying the chapter files of the documents in tdf/doc did not
trigger a rebuild when make was run again because these files were
not listed as dependencies. This commit adds the variable DOC_DPADD
to tendra.doc.mk to contain additional dependencies for a
documentation build. It also adds the chapter filenames to this
variable in the makefiles for each document.