Add a Dockerfile and script to run it
Define a base ubuntu image with the relevant compiler and libraries to
build tendra, and because docker has a pretty steep learning curve
itself, add a shell script that will run it with some pretty handy
non-defaults (like removing the container when you're finished with it).

I've also added some basic instructions.