Remove TDF_DIAG4 for codeview output.
I believe this code was never present in the build for the 4.1.2 release. From the CVS comments in 80x86/sco/config.h:

> * Revision 1.24  1998/01/11  15:15:04  pwe
> * remove NEWDIAGS

which was relatively late in the history (a few months before release). So I think TDF_DIAG4 underwent a few changes, and this code was disabled rather than updated.

I looked at updating it for TDF_DIAG4, but the code as written relied on a TDF access member, and I think TDF_DIAG4 is abstracted from those. So I'm just removing this, as the author intended.