Issue #93: -temp PREFIX_TMP CLI Option Does Not Work
The CLI option parser is an incredible piece of engineering
and works like a charm. However, there was a small typo
in the configuration of the `-temp` option which caused
the user-specified _value_ of the PREFIX_TMP to be treated
as the environment variable to set when the options are
parsed. This little fix makes everything okay.

Closes #93